fix(security,docker): clear the 5 remaining New Code findings (Stage 6)
After Stage 5 the Quality Gate had 1 condition left: Security Rating C
required A. Issues panel filtered to new code surfaced exactly 5
findings; this commit addresses all of them so the gate flips green.

Major Vulnerabilities (the actual gate blockers) — S6586
---
frontend/Dockerfile L10 and frontend/Dockerfile.dev L14 both ran
`npm ci` without `--ignore-scripts`. A poisoned transitive dep can ship
a malicious preinstall / postinstall lifecycle hook that runs arbitrary
code inside the build image. Adding `--ignore-scripts` blocks those
hooks at the boundary — neither vite nor esbuild needs its postinstall
step inside our image, and the apps still build clean.

Minor Code Smells
---
NotificationController.list (Stage 4A leftover, S1905):
  The Math.clamp call was written as `(int) Math.clamp((long) limit, 1, 100)`
  while I was over-thinking the overloads. Java 21 ships
  `Math.clamp(long, int, int)` returning int, so `limit` widens
  implicitly and neither cast is needed:
    int n = Math.clamp(limit, 1, 100);
  Sonar flagged both casts as redundant (×2 findings).

backend/Dockerfile L30 — S6504:
  The OpenTelemetry agent was being pulled in via `wget`, with two
  follow-up `chmod` / `chown` calls and a `wget` package install. Sonar
  recommends `ADD` with an HTTPS URL for this exact case — one
  instruction, no extra apk package, `--chown=appuser:appuser
  --chmod=644` apply ownership / mode atomically. Removing `wget` from
  the apk install also trims one transitive runtime dep.

No production behaviour change anywhere. NotificationControllerTest
still 11/11 green.

+++ backend/src/main/java/com/finansportali/backend/controller/NotificationController.java
@@ -30,7 +30,9 @@ public class NotificationController {
     public List<NotificationDto> list(@RequestParam(defaultValue = "30") int limit) {
         String userId = userService.getCurrentUserId();
         if (userId == null) return List.of();
-        int n = (int) Math.clamp((long) limit, 1, 100);
+        // Math.clamp(long, int, int) returns int in Java 21 — `limit` widens
+        // implicitly, no explicit casts needed.
+        int n = Math.clamp(limit, 1, 100);
         return repo.findByUserIdOrderByCreatedAtDesc(userId, PageRequest.of(0, n))
                 .stream().map(NotificationDto::from).toList();
     }
